+++ main.go
@@ -16,14 +16,19 @@ func main() {
 		log.Println("Execution Time: ", time.Since(start))
 	}()
 	client := cycletls.Init()
-	response, err := client.Do("https://www.google.com", cycletls.Options{
+	// response, err := client.Do("http://stackoverflow.com/q/16784419/727643", cycletls.Options{
+  response, err := client.Do("http://httpbin.org/post", cycletls.Options{
+
 		Body:      "",
-		Ja3:       "771,4865-4866-4867-49195-49199-49196-49200-52393-52392-49171-49172-156-157-47-53,0-23-65281-10-11-35-16-5-13-18-51-45-43-27-17513,29-23-24,0",
-		UserAgent: "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/101.0.4951.54 Safari/537.36",
+		Ja3:       "771,4865-4867-4866-49195-49199-52393-52392-49196-49200-49162-49161-49171-49172-156-157-47-53,0-23-65281-10-11-35-16-5-34-51-43-13-45-28,29-23-24-25-256-257,0",
+		UserAgent: "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:103.0) Gecko/20100101 Firefox/103.0",
 	}, "GET")
 	if err != nil {
 		log.Print("Request Failed: " + err.Error())
 	}
-	log.Println(response.Status,)
+	log.Println(response.ResponseUrl,)
+
+  log.Println(response.Cookies,)
+
 
 }